% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{Account}
\alias{Account}
\title{AdSense Management API Objects 
Accesses AdSense publishers' inventory and generates performance reports.}
\usage{
Account(id = NULL, name = NULL, premium = NULL, subAccounts = NULL)
}
\arguments{
\item{id}{Unique identifier of this account}

\item{name}{Name of this account}

\item{premium}{Whether this account is premium}

\item{subAccounts}{Sub accounts of the this account}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:23:34
filename: /Users/mark/dev/R/autoGoogleAPI/googleadsensev13.auto/R/adsense_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
