% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdCode}
\alias{AdCode}
\title{AdCode Object}
\usage{
AdCode(adCode = NULL)
}
\arguments{
\item{adCode}{The ad code snippet}
}
\value{
AdCode object
}
\description{
AdCode Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

