% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_functions.R
\docType{package}
\name{adsense_googleAuthR}
\alias{adsense_googleAuthR}
\alias{adsense_googleAuthR-package}
\title{AdSense Management API
Accesses AdSense publishers' inventory and generates performance reports.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:33:14
filename: /Users/mark/dev/R/autoGoogleAPI/googleadsensev13.auto/R/adsense_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/adsense
\item https://www.googleapis.com/auth/adsense.readonly
}
}

