% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdUnit.contentAdsSettings}
\alias{AdUnit.contentAdsSettings}
\title{AdUnit.contentAdsSettings Object}
\usage{
AdUnit.contentAdsSettings(AdUnit.contentAdsSettings.backupOption = NULL,
  backupOption = NULL, size = NULL, type = NULL)
}
\arguments{
\item{AdUnit.contentAdsSettings.backupOption}{The \link{AdUnit.contentAdsSettings.backupOption} object or list of objects}

\item{backupOption}{The backup option to be used in instances where no ad is available}

\item{size}{Size of this ad unit}

\item{type}{Type of this ad unit}
}
\value{
AdUnit.contentAdsSettings object
}
\description{
AdUnit.contentAdsSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings specific to content ads (AFC) and highend mobile content ads (AFMC - deprecated).
}
\seealso{
Other AdUnit functions: \code{\link{AdUnit.contentAdsSettings.backupOption}},
  \code{\link{AdUnit.feedAdsSettings}},
  \code{\link{AdUnit.mobileContentAdsSettings}},
  \code{\link{AdUnit}}
}
