% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{AdUnits}
\alias{AdUnits}
\title{AdUnits Object}
\usage{
AdUnits(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The ad units returned in this list response}

\item{nextPageToken}{Continuation token used to page through ad units}
}
\value{
AdUnits object
}
\description{
AdUnits Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
