% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{Alert}
\alias{Alert}
\title{Alert Object}
\usage{
Alert(id = NULL, message = NULL, severity = NULL, type = NULL)
}
\arguments{
\item{id}{Unique identifier of this alert}

\item{message}{The localized alert message}

\item{severity}{Severity of this alert}

\item{type}{Type of this alert}
}
\value{
Alert object
}
\description{
Alert Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
