% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_functions.R
\name{customchannels.get}
\alias{customchannels.get}
\title{Get the specified custom channel from the specified ad client.}
\usage{
customchannels.get(adClientId, customChannelId)
}
\arguments{
\item{adClientId}{Ad client which contains the custom channel}

\item{customChannelId}{Custom channel to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/adsense
\item https://www.googleapis.com/auth/adsense.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/adsense, https://www.googleapis.com/auth/adsense.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/adsense/management/}{Google Documentation}
}
