% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{ReportingMetadataEntry}
\alias{ReportingMetadataEntry}
\title{ReportingMetadataEntry Object}
\usage{
ReportingMetadataEntry(compatibleDimensions = NULL,
  compatibleMetrics = NULL, id = NULL, requiredDimensions = NULL,
  requiredMetrics = NULL, supportedProducts = NULL)
}
\arguments{
\item{compatibleDimensions}{For metrics this is a list of dimension IDs which the metric is compatible with, for dimensions it is a list of compatibility groups the dimension belongs to}

\item{compatibleMetrics}{The names of the metrics the dimension or metric this reporting metadata entry describes is compatible with}

\item{id}{Unique identifier of this reporting metadata entry, corresponding to the name of the appropriate dimension or metric}

\item{requiredDimensions}{The names of the dimensions which the dimension or metric this reporting metadata entry describes requires to also be present in order for the report to be valid}

\item{requiredMetrics}{The names of the metrics which the dimension or metric this reporting metadata entry describes requires to also be present in order for the report to be valid}

\item{supportedProducts}{The codes of the projects supported by the dimension or metric this reporting metadata entry describes}
}
\value{
ReportingMetadataEntry object
}
\description{
ReportingMetadataEntry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

