% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsense_objects.R
\name{SavedReports}
\alias{SavedReports}
\title{SavedReports Object}
\usage{
SavedReports(etag = NULL, items = NULL, nextPageToken = NULL)
}
\arguments{
\item{etag}{ETag of this response for caching purposes}

\item{items}{The saved reports returned in this list response}

\item{nextPageToken}{Continuation token used to page through saved reports}
}
\value{
SavedReports object
}
\description{
SavedReports Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
